Declaring new as long int

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,7 +11,7 @@ int main(void)
 	int i;
 	long int first;
 	long int second;
-	int new;
+	long int new;
 	long int sum;
 
 	first = 1;
